Refactor chat room logic into hooks
